Put a mock-for-data into the core example
