+++ LeanTest.sln
@@ -7,8 +7,6 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core", "Source\Core\Core.cs
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mock", "Source\Mock\Mock.csproj", "{E412F59B-73AF-4071-9275-A341595FCD5B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Mock.Examples.MsTest", "Source\Mock.Examples.MsTest\Mock.Examples.MsTest.csproj", "{D5D202EB-D9A1-47A7-9844-70BD649D7E4C}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JSon", "Source\JSon\JSon.csproj", "{4D4EC696-9F1E-473F-9A92-06A95E835C59}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Build", "Build", "{62FABEA8-7550-483D-839C-C0D50444D674}"
@@ -54,10 +52,6 @@ Global
 		{E412F59B-73AF-4071-9275-A341595FCD5B}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E412F59B-73AF-4071-9275-A341595FCD5B}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E412F59B-73AF-4071-9275-A341595FCD5B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{D5D202EB-D9A1-47A7-9844-70BD649D7E4C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D5D202EB-D9A1-47A7-9844-70BD649D7E4C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D5D202EB-D9A1-47A7-9844-70BD649D7E4C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D5D202EB-D9A1-47A7-9844-70BD649D7E4C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{4D4EC696-9F1E-473F-9A92-06A95E835C59}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{4D4EC696-9F1E-473F-9A92-06A95E835C59}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4D4EC696-9F1E-473F-9A92-06A95E835C59}.Release|Any CPU.ActiveCfg = Release|Any CPU
