+++ Source/Core.Examples.L0Tests/Domain/MyData.cs
@@ -1,8 +1,8 @@
 namespace Core.Examples.L0Tests.Domain
 {
-    public class MyData
-    {
-        public int First { get; set; }
-        public int Second { get; set; }
-    }
+	public class MyData
+	{
+		public string Key { get; set; }
+		public int Age { get; set; }
+	}
 }
