--- Source/Mock.Examples.MsTest/Application/IMyExternalService.cs
@@ -1,7 +1,0 @@
-namespace Mock.Examples.MsTest.Application
-{
-	public interface IMyExternalService
-	{
-		int GetAge(string key);
-	}
-}
