--- Source/Mock.Examples.MsTest/Application/MyApplicationService.cs
@@ -1,16 +1,0 @@
-﻿namespace Mock.Examples.MsTest.Application
-{
-    public class MyApplicationService
-    {
-        private readonly IMyExternalService _myExternalService;
-
-        public MyApplicationService(IMyExternalService myExternalService)
-        {
-            _myExternalService = myExternalService;
-        }
-        public int GetAge(string key)
-        {
-            return _myExternalService.GetAge(key);
-        }
-    }
-}
